+++ webcli/src/features/contextmenu/ContextMenuProvider.tsx
@@ -1,0 +1,10 @@
+import ContextMenu from './ContextMenu';
+
+export default function ContextMenuProvider({ children }: { children: React.ReactNode }) {
+  return (
+    <>
+      {children}
+      <ContextMenu />
+    </>
+  );
+}
